requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile cookbook/examples/sql/requirements.in
#
altair==5.3.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   openai
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
blinker==1.8.2
    # via streamlit
cachetools==5.3.3
    # via streamlit
certifi==2024.6.2
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   streamlit
    #   typer
distro==1.9.0
    # via openai
exceptiongroup==1.2.1
    # via anyio
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via
    #   phidata
    #   streamlit
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via
    #   openai
    #   phidata
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.4
    # via
    #   altair
    #   pydeck
jsonschema==4.22.0
    # via altair
jsonschema-specifications==2023.12.1
    # via jsonschema
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
numpy==1.26.4
    # via
    #   altair
    #   pandas
    #   pgvector
    #   pyarrow
    #   pydeck
    #   streamlit
openai==1.33.0
    # via -r cookbook/examples/sql/requirements.in
packaging==24.0
    # via
    #   altair
    #   streamlit
pandas==2.2.2
    # via
    #   -r cookbook/examples/sql/requirements.in
    #   altair
    #   streamlit
pgvector==0.2.5
    # via -r cookbook/examples/sql/requirements.in
phidata==2.4.20
    # via -r cookbook/examples/sql/requirements.in
pillow==10.3.0
    # via streamlit
protobuf==4.25.3
    # via streamlit
psycopg[binary]==3.1.18
    # via -r cookbook/examples/sql/requirements.in
psycopg-binary==3.1.18
    # via psycopg
pyarrow==16.1.0
    # via streamlit
pydantic==2.7.3
    # via
    #   openai
    #   phidata
    #   pydantic-settings
pydantic-core==2.18.4
    # via pydantic
pydantic-settings==2.3.1
    # via phidata
pydeck==0.9.1
    # via streamlit
pygments==2.18.0
    # via rich
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via
    #   phidata
    #   pydantic-settings
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via phidata
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via streamlit
rich==13.7.1
    # via
    #   phidata
    #   streamlit
    #   typer
rpds-py==0.18.1
    # via
    #   jsonschema
    #   referencing
shellingham==1.5.4
    # via typer
simplejson==3.19.2
    # via -r cookbook/examples/sql/requirements.in
six==1.16.0
    # via python-dateutil
smmap==5.0.1
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
sqlalchemy==2.0.30
    # via -r cookbook/examples/sql/requirements.in
streamlit==1.35.0
    # via -r cookbook/examples/sql/requirements.in
tenacity==8.3.0
    # via streamlit
toml==0.10.2
    # via streamlit
tomli==2.0.1
    # via phidata
toolz==0.12.1
    # via altair
tornado==6.4.1
    # via streamlit
tqdm==4.66.4
    # via openai
typer==0.12.3
    # via phidata
typing-extensions==4.12.2
    # via
    #   altair
    #   anyio
    #   openai
    #   phidata
    #   psycopg
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   streamlit
    #   typer
tzdata==2024.1
    # via pandas
urllib3==1.26.18
    # via requests